analysis/requirements.txt
# Analysis Tool Requirements
datasets==2.14.0
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.66.1
plotly==5.17.0
jupyter==1.0.0






